requirements.txt
fastapi
python-whois
uvicorn
aiohttp
python-dotenv